refactor(react-components): primitives の barrel を撤廃して index.tsx を本体にする
17 個の primitive ディレクトリで以下を実施:
- xxx.tsx → index.tsx、xxx.stories.tsx → index.stories.tsx にリネーム
- 中身ゼロの barrel index.ts を削除
- create-affordance は JSX を持たないため index.ts に統一

兄弟 primitive への deep import (../button/button) は folder import (../button)
に整理。main の HeadingContext import も同様。

primitives/index.ts は HeadingContext を internal に保つため、heading 行のみ
named export (export { Heading, Section }) に切替。
