+++ packages/react-components/src/primitives/button/index.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Button } from "./button";
+import { Button } from ".";
 
 const meta = {
 	title: "UI/Button",
