--- packages/react-components/src/primitives/button/index.ts
@@ -1,1 +1,0 @@
-export * from "./button";
