--- packages/react-components/src/primitives/combo-box/index.ts
@@ -1,1 +1,0 @@
-export * from "./combo-box";
