+++ packages/react-components/src/primitives/create-affordance/index.ts
@@ -1,1 +1,7 @@
-export * from "./create-affordance";
+import { cn } from "../../libs/utils";
+
+export const createAffordanceClass = cn(
+	"border border-border border-dashed bg-transparent text-muted-fg outline-none transition-colors",
+	"hover:border-solid hover:bg-secondary hover:text-fg",
+	"focus-visible:border-solid focus-visible:bg-secondary focus-visible:text-fg focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-inset",
+);
