--- packages/react-components/src/primitives/drawer/index.ts
@@ -1,1 +1,0 @@
-export * from "./drawer";
