--- packages/react-components/src/primitives/error-alert/index.ts
@@ -1,1 +1,0 @@
-export * from "./error-alert";
