--- packages/react-components/src/primitives/gym-context/index.ts
@@ -1,1 +1,0 @@
-export * from "./gym-context";
