--- packages/react-components/src/primitives/heading/index.ts
@@ -1,1 +1,0 @@
-export { Heading, Section } from "./heading";
