+++ packages/react-components/src/primitives/index.ts
@@ -4,7 +4,7 @@ export * from "./create-affordance";
 export * from "./drawer";
 export * from "./error-alert";
 export * from "./gym-context";
-export * from "./heading";
+export { Heading, Section } from "./heading";
 export * from "./label";
 export * from "./link";
 export * from "./main";
