--- packages/react-components/src/primitives/label/index.ts
@@ -1,1 +1,0 @@
-export * from "./label";
