--- packages/react-components/src/primitives/link/index.ts
@@ -1,1 +1,0 @@
-export * from "./link";
