--- packages/react-components/src/primitives/main/index.ts
@@ -1,1 +1,0 @@
-export * from "./main";
