--- packages/react-components/src/primitives/menu/index.ts
@@ -1,1 +1,0 @@
-export * from "./menu";
