--- packages/react-components/src/primitives/number-field/index.ts
@@ -1,1 +1,0 @@
-export * from "./number-field";
