--- packages/react-components/src/primitives/popover/index.ts
@@ -1,1 +1,0 @@
-export * from "./popover";
