--- packages/react-components/src/primitives/scrollable/index.ts
@@ -1,1 +1,0 @@
-export * from "./scrollable";
