--- packages/react-components/src/primitives/tabs/index.ts
@@ -1,1 +1,0 @@
-export * from "./tabs";
