--- packages/react-components/src/primitives/text-field/index.ts
@@ -1,1 +1,0 @@
-export * from "./text-field";
